[Halpy-143] Add UFI and OLM checks
Adds a five-minute timed Offline Mode Check to validate that OFFLINE MODE is not set. If it is set, warns Oper and the system monitoring channel set in config.

Also adds a scheduled, automatic syncing of the Fact database with HalpyBOT on a weeklybasis.

Also  removes two unused configuration flags that are left over from cut content.